updated work

+++ store/models.py
@@ -1,6 +1,8 @@
 from django.db import models
 
 from django.contrib.auth.models import User
+
+
 # Create your models here.
 # ==================CUSTOMER=============================
 
@@ -18,7 +20,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
 	name = models.CharField(max_length=200)
-	price = models.FloatField()
+	price = models.DecimalField(max_digits=10, decimal_places=2)
 	digital = models.BooleanField(default=False,null=True, blank=True)
 	image = models.ImageField(null=True, blank=True)
